+++ rust/aura/src/main.rs
@@ -71,6 +71,7 @@ fn main() -> Result<(), Error> {
         SubCmd::Cache(c) if c.search.is_some() => cache::search(cachep, &c.search.unwrap())?,
         SubCmd::Cache(c) if c.backup.is_some() => cache::backup(fll, cachep, &c.backup.unwrap())?,
         SubCmd::Cache(c) if c.clean.is_some() => cache::clean(fll, cachep, c.clean.unwrap())?,
+        SubCmd::Cache(c) if c.invalid => cache::invalid(fll, &alpm, cachep)?,
         SubCmd::Cache(c) if c.list => cache::list(cachep)?,
         SubCmd::Cache(c) if c.refresh => cache::refresh(fll, &alpm, cachep)?,
         SubCmd::Cache(_) => unimplemented!(),